Extract SearchBar and TransactionsTable to their own components
